+++ packages/orq-rc/src/orq_ai_sdk/models/createchunkop.py
@@ -257,7 +257,7 @@ class CreateChunkResponseBody(BaseModel):
     updated_by_id: str
     r"""The unique identifier of the user who updated the chunk"""
 
-    id: Optional[str] = "chunk_01JQRXCBFHJ7ECK0QSR9X0Q1TF"
+    id: Optional[str] = "chunk_01JQS70Y0FJ7PHPRGRMRP0ZYKS"
     r"""Unique identifier for the element"""
 
     enabled: Optional[bool] = True