+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JQRXCBFEGK9VA8DNVBQRKHN5"
+        "01JQS70Y0B4TXQJK29Q28XD6BZ"
     )
     r"""The id of the resource"""