+++ packages/orq-rc/RELEASES.md
@@ -1029,3 +1029,13 @@ Based on:
 - [python v3.3.0-rc.21] packages/orq-rc
 ### Releases
 - [PyPI v3.3.0-rc.21] https://pypi.org/project/orq-ai-sdk/3.3.0-rc.21 - packages/orq-rc
+
+## 2025-04-01 17:41:34
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.527.0 (2.564.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.3.0-rc.22] packages/orq-rc
+### Releases
+- [PyPI v3.3.0-rc.22] https://pypi.org/project/orq-ai-sdk/3.3.0-rc.22 - packages/orq-rc